Split shell queue admission tests
